Probing has strike abort safety in it.
This is because a strike elsewhere on the bomb, has a 20% chance of
causing probing to change its frequency, and thus make the submitted
answer an additional strike.   Also, if that strike happens in the
middle of the cycle command,  the frequency change if it happens will
make that set of frequencies read unsuable.