!bomb timestamp is now possible, !bomb edgework expanded
Some modules depend on knowing the day and possibly the time the bomb
has started, such as cheap checkout.  Figured it would be best to get
something that responds into chat the exact moment the bomb was spawned.

Additionnaly, edgework has been expanded

!bomb edgework - Goes through ALL edgework
!bomb edgework [face] - Views the specific face of the bomb, in case for
example, you noted that serial number was on the top face, but didn't
quite get all of it.