Fix for time-critical interactions. Commits 74a2618d24562f2e6632308203c55ec9c43ca9bb and 90273bc4c64efff38f38106b102c2faba89766c8 introduced a flaw whereby time-critical mod-side component solvers that "don't do the right thing" would end up executing incorrectly. Example modules of this would be Astrology, Colour Flash, etc. There was a flag to indicate 'delayed invokation', but this is really not the right way to solve this issue, as that would mean that TP would have to be updated every time a time-critical module is released with TP support.
The definition of "don't do the right thing" as mentioned in the commit descriptions of the referenced commits above describes that a solver *must* yield return something before doing any sort of interaction, if the parsed command was valid. While that is true of internal interaction code, the same is/was not particularly true for external support. There was no documentation that explicitly stated that the end-developer must yield return something before doing absolutely anything else, and also this seemed kludgy in its own right in the internal implementation anyway.

I have reinstated some of the old code whereby the forcible yield return before executing the .MoveNext() on the returned enumerator is back for all cases of external module support. The side-effect of this is that command parse errors are not handled ideally, causing the artefact of the module being focused as if it did receive a valid command, even if under the hood the command parse fails and does nothing.

The ideal implementation would fix both problems (which might be an impossibility without forcing developers to have to change existing code to suit), but this solution I feel is a far more desirable result, as it therefore allows all modules to work properly without issue, just with the slight issue that the bomb looks like it receives commands that it shouldn't.