Module data loading
Now the module data of twitch play supported modules can be updated
live, without needing to rebuild Twitch Plays, or even rebooting the
game.   The data can now reside in ModuleInformation.json,  which
contains all the properties of a given module.

validCommands is a list of regular expression matching strings.  If
provided the command to that module will be validated against that list,
and if there isn't at least one command that it passes, command
invokation does not happen.

The other option, DoesTheRightThing, if set, Tells twitch plays to
instantly hand over control to the module,  and let the module decide if
the command supplied was valid or not,  and process the command if it
is.  DoesTheRightThing is a white list based boolean, which is false by
default.

Also, the validCommands list will always overrule DoesTheRightThing, if
both are present.

Finally,  if an error is spotted in the module help text, or the !<id>
manual command produces a broken link,  both of these can be fixed for
the next loaded bomb,  just by modifying the json file now.